PCB: Convert to KiCad 9
This is the result of opening the design in KiCad 9, going through the
schematic and board setup dialogs and saving again. This updates to new
file extensions and severely changes the internal file contents, but
should not modify the design in any way.

This also removes the PCB-cache.lib file, since it is no longer needed
since KiCad 6.