PCB: Update Feather symbol and footprint
This changes from a locally created one to the KiCad library version of
it (which has seen some changes before it was merged).

This also seems to completely rewrite the .kicad_pcb file, so maybe the
earlier conversion to KiCad9 did not work (or was not committed), so
the diff of this commit on the PCB is useless.